format_x out of ParameterSet
